+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mobile_translation_app/pages/my_home_page.dart';
+import 'package:mobile_translation_app/my_home_page.dart';
 
 void main() {
   runApp(const MyApp());
